Remove header element, add imgs and modify classes
Removed the header element from section two as I decided to use that
section to display images and decided a header for this was redundant.

I added four imgs to section two and gave one of these elements an
additional class of "yellow" to select it independently in the css and
thus adjust its size independently from the other img elements. I did
this because, I was displaying the imgs within a parent div with the
css attribute "overflow: hidden" on the parent, which means only a
portion of the image was to be displayed. This was intended effect.
However, the portion of this image being displayed for this element
was too "zoomed in". So, by specifically selecting and then reducing the
width of this img I was able to display a fuller scope of it within the
parent element. This looks much nicer.

Additionally, I corrected the flex-container element with the class of
"two" and replaced it with the class "three" instead, as this was the
third flex-container and not the second. I then assigned the "two" class
to it's appropriate element, the second flex-container.