Add and modify classes, style imgs
I added new classes such as the "three" class which allows me to
independently target and resize the container element for the third
section. Additionally, I modified the attritubes of the "two" class used
for the container element for the second section, to appropriately size
it so that section one and section two combine to take up the full
height of the screen. I edited the size of the subsequent containers and
flex-containers for the subsequent sections in order that any two
subsequent container elements will always combine to take up 100% of the
screen height at any given point. This looks nice and gives a uniformed
and orderly appearance to the page as a whole.

Additionally, I styled the 4 img elements in section two so that they
are nested within a parent container that has the attritube
"overflow: hidden". This allows only a portion of the larger image to be
shown in porportion to the size of it's parent. Due to the large size of
the images in  question, I feel this approach looks much nicer than
simply resizing the images to fit the size of the parent as with that
approach, in this case, you lose a lot of scale. I then added a drop
shadow behind these imgs to make them pop. And gave the img in section
one a border radius of 12px to give it a more rounded appearance in
accordance with the section two imgs.