Add and modify class selectors
Added class selectors for "red" which can be used aside the existing
badge class to produce a badge element with a red background. Added
class selectors for testimonial which gives this container class a grey
background and rounded edges to serve as a card-style element. Added
class selector for card-text which is used for text within the
testimonial card and stylized the font to be grey and italicized. This
styling indicates that the text is quoting somebody. I used a type
selector to target the span element that is a child of the card-text
class element and give this text a bolder font-weight, a black color and
normal font style. This is to indicate that this text is not a quote but
the name of who has been quoted. I also changed the footer color to
black and the section four container to a dark grey color that matches
the background color of section one. These two background color changes
provide a nice visual contrast to preceeding elements.